+++ src/grpc/client/js/dodex/index.js
@@ -6,28 +6,29 @@ import "jsoneditor/dist/jsoneditor.min.css"
 import jsonEditor from "jsoneditor"
 import "@fortawesome/fontawesome-free/js/all.js";
 import "@fortawesome/fontawesome-free/js/fontawesome.js";
+import "../../css/dodex_handicap.css";
 window.JSONEditor = jsonEditor;
 
 if (document.querySelector(".top--dodex") === null) {
     // Content for cards A-Z and static card
-    dodex.setContentFile("handicap/content.js");
+    dodex.setContentFile("handicap/handicap_info.js");
     const server = window.location.hostname + (window.location.port.length > 0 ? ":" + window.location.port : "");
 //    const server = "127.0.0.1:8089"; // (8089)-development (8080)-production
     dodex.init({
       width: 375,
-      height: 200,
+      height: 220,
       left: "50%",
       top: "100px",
       input: input,
-      private: "full",
+      private: "partial",
       replace: true,
       mess: mess,
       server: server
     }).then(function () {
       // Add in app/personal cards
-      for (let i = 0; i < 4; i++) {
-        dodex.addCard(getAdditionalContent());
-      }
+//      for (let i = 0; i < 4; i++) {
+//        dodex.addCard(getAdditionalContent());
+//      }
       /* Auto display of widget */
       // dodex.openDodex();
     });