+++ src/main/java/dmo/fs/db/reactive/DbReactiveSqlBase.java
@@ -20,7 +20,7 @@ import org.jooq.impl.DSL;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.websocket.Session;
+import jakarta.websocket.Session;
 import java.sql.Timestamp;
 import java.text.DateFormat;
 import java.time.OffsetDateTime;