+++ src/main/java/dmo/fs/db/reactive/DodexReactiveRouter.java
@@ -18,7 +18,7 @@ import io.vertx.reactivex.core.shareddata.SharedData;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.websocket.Session;
+import jakarta.websocket.Session;
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;