+++ src/main/java/dmo/fs/kafka/KafkaEmitterDodex.java
@@ -7,8 +7,8 @@ import java.util.TimeZone;
 import java.util.concurrent.CompletableFuture;
 
 import javax.annotation.Priority;
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;